Refactors Sidebars out of SidebarLayout
